Add ActiveRecord introspection and association analysis utilities
- Introduced ActiverecordIntrospection module for checking ActiveRecord availability and eager loading models.
- Added AssociationScopeFiltering module to filter associations based on session context.
- Implemented AssociationExtractor service for extracting and analyzing ActiveRecord model associations.
- Created DatasetBuilder service to transform model association data into a standardized dataset format.
- Developed ModelDiscovery service for discovering and filtering ActiveRecord models relevant for analysis.
- Refactored ModelAssociationAnalyzer to utilize the new services and concerns for improved organization and functionality.
